requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
psycopg[binary,pool]==3.2.9
SQLAlchemy==2.0.32
pydantic==2.8.2
langchain==0.2.11
langchain-community==0.2.10
langchain-postgres==0.0.11
sentence-transformers==3.0.1
numpy==1.26.4
python-dotenv==1.0.1
langchain-openai==0.1.16